requirements.txt
accelerate==0.22.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auto-gptq==0.4.2
bitsandbytes==0.42.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.15.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
en-core-web-lg==3.7.1
en-core-web-md==3.7.1
en-core-web-sm==3.7.1
en-core-web-trf==3.7.3
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.15.4
FlagEmbedding==1.2.11
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langcodes==3.4.0
langsmith==0.1.141
language_data==1.2.0
lightfm==1.17
loguru==0.7.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.54.3
orjson==3.10.11
packaging==24.1
pandas==2.0.3
parso==0.8.4
peft==0.6.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.6.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
rdflib==7.0.0
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
rouge==1.0.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==72.1.0
shtab==1.7.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.2
spacy-alignments==0.9.1
spacy-curated-transformers==0.2.2
spacy_dbpedia_spotlight==0.2.6
spacy-entity-linker==1.0.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
spacyfishing==0.1.8
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.2
sympy==1.13.2
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.15.2
torch==2.1.1
torchvision==0.16.1
tornado==6.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.36.2
triton==2.1.0
trl==0.7.11
typer==0.9.4
typing_extensions==4.12.2
typing-inspect==0.9.0
tyro==0.8.10
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
wheel==0.43.0
xxhash==3.5.0
yarl==1.9.7
